feat(modularization): S-5 extract kanban DnD to web/modules/kanban-dnd.js
Extracts the full drag-and-drop layer (~258 lines) from app.js into a dedicated
ES module. Both HTML5 drag/drop (desktop) and touch drag (mobile) are covered.

Design:
- kanban-dnd.js imports loadKanban from kanban.js (calls it after /move) and
  showToast + t from their respective modules.
- No kanbanState reads: DnD only POSTs /api/kanban/:id/move and calls
  loadKanban() -- the board re-renders through the existing hook chain.
- Re-render path: loadKanban() -> renderKanban() -> window._onKanbanRefresh?.()
  (Gantt). No extra wiring needed.
- wireKanbanColumnDnD and wireKanbanCardTouchDnD exported; app.js imports them
  and passes to initKanban() unchanged.

2868/2868 tests pass. /modules/kanban-dnd.js -> 200 application/javascript.
